# Optimized writing of byte array to parquet (#1764) (2x faster)